refactor: extract TranscriptionModels.swift (#648 slice A)
First slice of #648 (TranscriptionClient.swift decomposition). Extracts 3
model structs (TranscriptionRequest, TranscriptionResult, TranscriptionSegment)
from TranscriptionClient.swift into a new TranscriptionModels.swift.

Byte-identical move; SHA-256 32c99bb5a956c79d4e65f472f157ec719ff201b5933c2773e8a6503f6a8d3b06
verified against origin/main:TranscriptionClient.swift lines 4-43.

Reduces TranscriptionClient.swift from 753 lines to 711 lines.

Tests: 667 (unchanged).
